// wls-gui-wahllokalsystem/src/api/wls-clients/generated-vorfaelleundvorkommnisse-api/api.ts
export interface EreignisDTO {
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'beschreibung'?: string;
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'uhrzeit'?: string;
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'ereignisart': EreignisDTOEreignisartEnum;
}

export const EreignisDTOEreignisartEnum = {
    Vorfall: 'VORFALL',
    Vorkommnis: 'VORKOMMNIS'
} as const;

export type EreignisDTOEreignisartEnum = typeof EreignisDTOEreignisartEnum[keyof typeof EreignisDTOEreignisartEnum];
// ...
export interface WahlbezirkEreignisseDTO {
    /**
     * 
     * @type {string}
     * @memberof WahlbezirkEreignisseDTO
     */
    'wahlbezirkID': string;
    /**
     * 
     * @type {boolean}
     * @memberof WahlbezirkEreignisseDTO
     */
    'keineVorfaelle'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof WahlbezirkEreignisseDTO
     */
    'keineVorkommnisse'?: boolean;
    /**
     * 
     * @type {Array<EreignisDTO>}
     * @memberof WahlbezirkEreignisseDTO
     */
    'ereigniseintraege'?: Array<EreignisDTO>;
}